Components: Add a ReduxFormFieldset component
Wrapper around a FormFieldset that can display a label, validation error and explanation.
It can be parametrized by the type of input component it should render.

I'm actually moving an existing component from `SimplePaymentsDialog` to `components/redux-forms`
rathe that creating it from scratch.
